requirements.txt
absl-py==1.3.0
asttokens==2.2.1
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
boto3==1.26.41
botocore==1.29.41
bs4==0.0.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
filelock==3.9.0
fire==0.5.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
ipython==8.7.0
jedi==0.18.2
jmespath==1.0.1
keyboard==0.13.5
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.23.3
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.5
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
sahi==0.11.9
scipy==1.9.3
seaborn==0.12.1
selenium==4.7.2
shapely==2.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wget==3.2
wincertstore==0.2
wsproto==1.2.0
yolov5==7.0.6
yolov7==0.0.1
